requirements-dev.txt
# NEEDS UPDATING
mock
moto>=0.3.1
nose
django_nose
nose-testconfig
selenium
pep8==1.5.0
pyflakes

